+++ app/src/main/java/uk/co/jakelee/blacksmith/main/FurnaceActivity.java
@@ -69,18 +69,18 @@ public class FurnaceActivity extends Activity {
         }
 
         // Display item name and description
-        displayItemInfo((int) mViewFlipper.getCurrentView().getTag(), 1);
+        displayItemInfo((Long) mViewFlipper.getCurrentView().getTag(), 1);
 
         // Display item ingredients
         TableLayout ingredientsTable = (TableLayout) findViewById(R.id.ingredientsTable);
-        dh.createItemIngredientsTable((int) mViewFlipper.getCurrentView().getTag(), 1, ingredientsTable);
+        dh.createItemIngredientsTable((Long) mViewFlipper.getCurrentView().getTag(), 1, ingredientsTable);
     }
 
     public void closeFurnace(View view) {
         finish();
     }
 
-    public void displayItemInfo(int itemId, int state) {
+    public void displayItemInfo(Long itemId, int state) {
         View furnace = findViewById(R.id.furnace);
         Item item = dbh.getItem(itemId);
         Inventory count = dbh.getInventory(itemId, state);
@@ -101,7 +101,7 @@ public class FurnaceActivity extends Activity {
     }
 
     public void smelt1(View v) {
-        int itemId = (int) mViewFlipper.getCurrentView().getTag();
+        Long itemId = (Long) mViewFlipper.getCurrentView().getTag();
 
         Item item = dbh.getItem(itemId);
         if (dbh.createItem(itemId, 1, 1, 1)) {
@@ -126,10 +126,10 @@ public class FurnaceActivity extends Activity {
                 mViewFlipper.showPrevious();
             }
 
-            displayItemInfo((int) mViewFlipper.getCurrentView().getTag(), 1);
+            displayItemInfo((Long) mViewFlipper.getCurrentView().getTag(), 1);
 
             TableLayout ingredientsTable = (TableLayout) findViewById(R.id.ingredientsTable);
-            dh.createItemIngredientsTable((int) mViewFlipper.getCurrentView().getTag(), 1, ingredientsTable);
+            dh.createItemIngredientsTable((Long) mViewFlipper.getCurrentView().getTag(), 1, ingredientsTable);
 
             return super.onFling(e1, e2, velocityX, velocityY);
         }
